MCPP/Filemanager-Tool/requirements.txt
mcp
python-dotenv
google-generativeai   # for Gemini API
